==> Credit risk modelling use case/notebooks/01_data_pre_processing.Rmd <==
---
title: "Data Preprocessing"
author: "41084.1_XAI_for_Finance"
output: html_document
---

This notebook contains the project's outputs with respect to WP4, specifically looking at data pre-processing. The script provides an automated approach for running the necessary pre-processing for building various ML models. 

```{r}
# ---------------------------------- Factorize -----------------------------------
#' 
#' factorize data.frame
#'
#' Convert character columns in a data.frame to factors.
#'
#' Given a data.frame, convert character columns to factors.  By default,
#' all character columns are converted, but the user can specify which columns
#' to convert or to not convert.
#'
#' @param df [data.frame] input data frame
#' @param only [vector] of column names. Only convert these columns to factors.
#' @param ignore [vector]  of column names. Do not convert these columns.
#' @param verbose [boolean] Print information about the process
#' 
#' @return A data.frame with all character columns (or only a subset of them
#'   if the parameters were set) converted to factors.
#'   
#' @example
#'  df <- data.frame(a = 1:3, b = letters[1:3], c = LETTERS[1:3],
#'                  stringsAsFactors = FALSE)
#' str(df)
#' str(dfFactorize(df))
#' 
#' 
dfFactorize <- function(df, only = c(), ignore = c(), verbose = TRUE) {
  # Check parameters
  stopifnot(
    is.data.frame(df),
    length(only) == 0 || length(ignore) == 0,
    all(c(only, ignore) %in% colnames(df))
  )
  
  # Determine which columns to change
  colsConvert <- rep(TRUE, ncol(df))
  if (length(only) > 0) {
    colsConvert <- colnames(df) %in% only
  }
  if (length(ignore) > 0) {
    colsConvert <- !(colnames(df) %in% ignore)
  }
  colsConvert <- colsConvert & sapply(df, is.character)
  
  if (verbose){
    cat("converting ", length(which(colsConvert)), "chatacter columns to factor columns")
  }
  
  # Convert specific columns to factors
  df[colsConvert] <- lapply(df[colsConvert], as.factor)
  
  return(df)
}


# ---------------------------------------------- Binarize ------------------------------------------
#' 
#' Function to binarize a mutliclass dataset 
#' 
#' @param df [data.frame] Dataset to binarize of type data.frame
#' @param target [integer/string] Index or name of the column containing the target variable
#' @param level1 [vector] vector containing all levels of the feature to be set to 1
#'                        everything else will be set to 0
#' @param verbose [boolean] Print information about the process
#' 
#' @example
#' 
#' df_bin = binarize(df, target = "loan_status", level1 = "Default")
#' 
#' 
binarize <- function(df, target, level1, verbose = TRUE){
  stopifnot(
    is.data.frame(df),
    target %in% colnames(df),
    any(level1 %in% df[, target])
  )
  for (i in 1:length(level1)){
    if (length(grep(level1[i], df[,target])) < 1){
      warning(paste(level1[i], "is not a level of", target, "and will not be used"))
      level1 = level1[-i]
    }
  }
  stopifnot(length(level1) > 0)
  
  if (verbose){
    cat("converting:\n\n", paste(level1, collapse = "\n"), "\n\nto Level 1\n")
    cat("------- and --------- \n", 
        paste(setdiff(levels(as.factor(df[, target])), level1), collapse = "\n"), 
        "\n\n to Level 0\n")
  }
  
  t_vec = as.character(df[, target])
  t_vec[!(t_vec %in% level1)] = 0
  t_vec[t_vec %in% level1] = 1
  df[, target] = as.factor(t_vec)
  
  if (verbose){
    cat("--- New Target distribution ---\n")
    print(table(df[, target]))
  }
  
  return(df)
}


# -------------- undersampling for binary target ----------------
#' 
#' Undersampling of a data frame with a binary target
#'
#' @param df [data.frame] input data frame
#' @param target [string] column name of the target feature
#' @param verbose [boolean] Print information about the process
#'
#' @return balanced data.frame
#' 
#' @example 
#' 
#' undersample(df, "loan_status")
#'
undersample <-function(df, target, verbose = TRUE){
  stopifnot(
    is.data.frame(df),
    target %in% colnames(df),
    length(table(df[, target])) == 2
  )
  
  tbl = table(df[, target])
  
  if (verbose){
    cat("------------ input distribution of target ------------ \n")
    print(tbl)
  }
  
  over_class = which.max(tbl)
  under_class = which.min(tbl)
  
  bool_over = names(tbl[over_class]) == df[, target]
  idx_over = 1:length(bool_over)
  idx_down = sample(idx_over[bool_over], size = tbl[under_class])
  
  bool_under = names(tbl[under_class]) == df[, target]
  idx_under = 1:length(bool_under)
  idx_under = idx_under[bool_under]
  
  idx = sort(c(idx_down, idx_under))
  
  balanced = df[idx, ]
  
  if (verbose){
    cat("------------ output distribution of target ------------ \n")
    print(table(balanced[, target]))
  }
  
  return(balanced)
}

# ---------------------------- Ad missing Features for numerical variables -----------------
#'
#' Function adding additional columns to a data frame
#' for all numerical columns with at least one missing value
#' containing information about whether the value is missing 
#' in the original column. 
#' These additional columns are named "is_missing" followed by the original name
#'
#'
#' @param df [data.frame] input data frame
#' @param only [vector] of column names. Only apply to these columns
#' @param ignore [vector]  of column names. Do not apply to these columns.
#' @param verbose [boolean] Print information about the process
#'
#' @example
#'
#' add_missing_args(df, only = c("loan_amnt"))
#'
add_missing_args <- function(df, only = c(), ignore = c(), verbose = TRUE, out.dist = FALSE){
  # Check parameters
  stopifnot(
    is.data.frame(df),
    length(only) == 0 || length(ignore) == 0,
    all(c(only, ignore) %in% colnames(df))
  )
  
  # Determine which columns to change
  colsConvert <- rep(TRUE, ncol(df))
  if (length(only) > 0) {
    colsConvert <- colnames(df) %in% only
  }

  if (length(ignore) > 0) {
    colsConvert <- !(colnames(df) %in% ignore)
  }

  colsConvert <- colsConvert & sapply(df, is.numeric) & sapply(df, function(x) any(is.na(x)))
  
  new_names = paste0("is_missing_", names(colsConvert[which(colsConvert)]))
  
  if (verbose){
    cat("Found ", length(new_names), "numerical columns containing missing values. \n
        Creating additianoal features for following colums: \n",
        paste(names(colsConvert[which(colsConvert)]), 
              collapse = "\n"), "\n")
  }
  
  na_df = replace(df[, colsConvert], !is.na(df[, colsConvert]), 0)
  na_df = replace(na_df, is.na(na_df), 1)
  
  
  colnames(na_df) = new_names
  df = cbind(df, na_df)
  
  if (verbose){
    cat("--------------------------- Distribution of new features --------------------------- \n")
    cat("(set `verbose=FALSE` to prevent calculation of distributions) \n")
    distr = sapply(na_df[, new_names], table)
    print(distr)
    return(list("data" = df, "distributions" = distr))
  }
  
  return(list("data" = df, "distributions" = "set `verbose=TRUE` to recive distributions"))
}



# ------------------------------------ Transformation ---------------------------------------
#'
#'  Powertransform all numerical positive variables using boxcox transformation 
#'  
#'  @requires library(EnvStats)
#'  
#'  @param df [data.frame] input data frame
#'  @param lambda	[integer] finite numeric scalar indicating what power to use for the Box-Cox transformation.
#'  @param verbose [boolean] Print information about the process
#' 
#'  @example 
#'  
#'  power_transform(df, lambda = 0.7)
#' 
power_transform <- function(df, lambda = 0.5, verbose = TRUE){
  stopifnot(is.data.frame(df))
  
  suppressWarnings({
    mask_pos = sapply(df, function(x) !any(x <= 0, na.rm = TRUE) & is.numeric(x))
  })
  
  stopifnot(sum(mask_pos) > 0)
  
  if (verbose){
    cat("Transforming ", sum(mask_pos), "columns: \n", 
        paste(colnames(df)[mask_pos], collapse = "\n"))
  }
  

  ret = lapply(df[, mask_pos], function(x) boxcoxTransform(x, lambda = 0.5))
  df[, mask_pos] = as.data.frame(ret)
  
  return(df)
}



# ----------------------------- Grouping levels of feature ---------------------
#' 
#' Function to group attributes containing free text based on defined key words
#' Those keywords are given as a vector where earlier keywords have a higher
#' priority than later keywords. For example; if the `keywords` vector is 
#' c("Manager", "operations") ant the text in the observation is "Operations Manager" 
#' the new level will be "manager".
#' 
#'  @param df [data.frame] input data frame
#'  @param feature [string] defining the name of the feature to use
#'  @param keywords [vector] defining the keywords to relevel
#'  @param replace_original [boolean] replacing the original feature of add a new one 
#'  @param verbose [boolean] Print information about the process
#' 
#'  @example 
#' 
#' df_gr = group_text_feature(df,
#'                            feature = "emp_title",
#'                            keywords = c("analyst", "manager"))
#' 
group_text_feature <- function(df, feature, keywords, 
                               replace_original = FALSE,
                               verbose = TRUE){
  # Check parameters
  stopifnot(
    is.data.frame(df),
    any(feature %in% colnames(df)),
    !is.numeric(df[, feature])
  )
  
  if (verbose){
    cat("grouping ", feature, "to following levels: \n", paste(keywords, collapse = "\n"))
  }

  keywords = rev(keywords)
  lvl_vec = rep(NA, length(df[, feature]))

  for (level in keywords){
    lvl_vec[grepl(tolower(level), tolower(df[, feature]))] = level
  }

  lvl_vec = replace(lvl_vec, is.na(lvl_vec), "other")
  lvl_vec = as.factor(lvl_vec)
  
  if (replace_original){
    if (verbose){
      cat("\n Replacing original feature in df")
    }
    df[, feature] = lvl_vec
  }
  if (!replace_original){
    if (verbose){
      cat("\n Creating new feature in df")
    }
    oldnames = colnames(df)
    df = cbind(df, lvl_vec)
    colnames(df) = c(oldnames, paste0(feature, "_grouped"))
  }
  if (verbose){
    cat("\new feature distribution: \n")
    print(table(lvl_vec))
  }
  
  return(df)
}


################################## Import data and apply functions ##################################
accepted_2007_to_2018Q4 <- read.csv("C:/Users/obri/OneDrive/IDP/Projects/_03_XAI/accepted_2007_to_2018Q4.csv", header=TRUE)
df = accepted_2007_to_2018Q4

# ----------------------------------- Show target distribution -----------------
x0 = sum(df$loan_status != "Default" & df$loan_status != "Charged Off")
x1 = sum(df$loan_status == "Default" | df$loan_status == "Charged Off")
xx = barplot(c(x0, x1), names.arg = c("0", "1 Charged Off"))
text(x = xx, y = c(x0, x1)-1e5, labels = c(x0, x1))


# ----------------------------- manually apply functions -----------------------
# library(EnvStats)
# df_fact <- dfFactorize(df, ignore = c("id", "member_id"))
# df_bin <- binarize(df_fact, target = "loan_status", level1 = c("Charged Off", "Default"), verbose = T)
# df_bl <- undersample(df_bin, target = "loan_status")
# df_na <- add_missing_args(df_bl)$data
# df_transformed <- power_transform(df_na)
# df_gr = group_text_feature(df,
#                         feature = "emp_title",
#                         keywords = c("CEO", "CFO", "CMO", "manager", 
#                                      "supervisor", "engineer", "professor", "sales", 
#                                      "teacher")
# )


# ----------------------------------- main -------------------------------------
library(EnvStats)
#' Apply pre processing steps to a data frame
#' 
#' 
#' @requires library(EnvStats)
#'
#' @param df [data.frame] input data frame
#' @param used_methods [vector] of methods to be applied.
#' @param exclude [vector]  of `used_methods` not to be applied.
#' @param trgt [string] column name of the target feature
#' @param target.level.1 [vector] vector containing all levels of `trgt` to be set to 1
#'                        everything else will be set to 0
#' @param verbose [boolean] Print information about the process
#'
#' @example
#'
#' clean_df <-  main(df, trgt = "loan_status", target.level.1 = c("Charged Off", "Default"))
#' 
#' 
main <- function(df, 
                 used_methods = c("factorize", 
                                  "binarize", 
                                  "balance", 
                                  "na_handling",
                                  "transform",
                                  "group_text",
                                  "remove_na_cols"),
                 exclude = c(),
                 verbose = TRUE,
                 trgt = "loan_status",
                 target.level.1 = c("Charged Off", "Default")
                 ){
  stopifnot(
    is.data.frame(df),
    trgt %in% colnames(df),
    any(target.level.1 %in% df[, trgt])
  )
  
  if (length(exclude) > 0) {
    used_methods = setdiff(used_methods, exclude)
  }
  
  if (verbose){
    cat("\n\n ------  Start ------ \n")
    cat("Pre-processing of data frame containing ", 
        nrow(df), " observations of ", ncol(df), "variables \n ---- \n")
    cat("following methods will be applied: \n", paste(used_methods, collapse = "\n"), "\n")
    t.start = Sys.time()
  }

  if ("factorize" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Factorize ------ \n")
    }
    df = dfFactorize(df, verbose = verbose, ignore = c("id", "member_id"))
  }
  
  if ("binarize" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Binarize ------ \n")
    }
    df = binarize(df, target = trgt, level1 = target.level.1, verbose = verbose)
  }
  
  if ("balance" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Balance ------ \n")
    }
    df = undersample(df, target = trgt)
    
  }
  
  if ("na_handling" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Handling missing values ------ \n")
    }
    df <- add_missing_args(df)
    df <- df$data
  }
  
  if ("transform" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Feature transformation ------ \n")
    }
    df <- power_transform(df)
  }
  
  if ("group_text" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Grouping text fratures ------ \n")
    }
    df = group_text_feature(df,
                             feature = "emp_title",
                             keywords = c("CEO", "CFO", "CMO", "manager", 
                                          "supervisor", "engineer", "professor", "sales", 
                                          "teacher")
                            )
  }
  
  if ("remove_na_cols" %in% used_methods){
    if (verbose){
      cat("\n\n ------ Grouping text fratures ------ \n")
    }
    df = df[,which(unlist(lapply(df, function(x) !all(is.na(x)))))]
  }
  
  if (verbose){
    cat("\n\n ------ Finished Pre - Processing ------ \n")
    cat("Conversion took:\n")
    print(Sys.time() - t.start)
    cat("\n --------------- \n")
    cat("\nreturning data frame containing ", nrow(df), " observations of ", ncol(df), "variables \n ---- \n")
    cat("with following structure:\n")
    print(str(df))
  }
  
  return(df)
}

clean_df <- main(df, trgt = "loan_status", target.level.1 = c("Charged Off", "Default"))

# save processed file
write.csv(clean_df, file = "accepted_2007_clean.csv")


```
