==> Credit risk modelling use case/notebooks/03_further_feature_selection.Rmd <==
---
title: "03_further_feature_selection"
author: "41084.1_XAI_for_Finance"
output: html_document
---


This notebook contains the project's outputs with respect to WP4, specifically looking at feature selection. The script provides a sensible approach for selecting the features to be included in a credit risk use case looking exclusively at personal lending.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

libraries = c("readr", "corrplot", "RColorBrewer", "tidyverse", "kernlab", "knitr", "tidytable", "MLmetrics", "stargazer", "dplyr", "purrr", "xtable", "base", "arulesViz", "gbm", "ggplot2", "DescTools","stylo", "igraph", "xgboost", "DALEX", "DALEXtra", "stats", "ranger", "ALEPlot", "party", "lime", "MASS", "iml", "ROCR", "rpart", "e1071", "shapper", "caret","MLmetrics", "Matrix", "Boruta", "Metrics", "fastshap", "shapper", "fscaret", "kknn", "knitr", "reshape", "reshape2", "shapper", "randomForest", "Hmisc", "networkD3", "emstreeR", "readxl", "pdp", "pROC", "randomForestExplainer", "logistf")

lapply(libraries, function(x) if (!(x %in% installed.packages())) {
  install.packages(x)
})

lapply(libraries, library, quietly = TRUE, character.only = TRUE)

rm(list=ls())
options(warn = -1)
set.seed(7)
```

```{r}
##### Load Data #####
df <- readRDS("~/GitHub/IWA_xai/2022_repo/Loan Performance Use Case/data/clean_df_after_preliminary_fs.rds")
```

```{r}
##### Correlation check #####
df_numeric <- select_if(df, is.numeric)
cor_numeric <- cor(df_numeric)
index <- which(abs(cor_numeric) > .90 & abs(cor_numeric) < 1, # your criteria
               arr.ind = T) # the result of the which function is now in rows & columns
cbind.data.frame(variable_1 = rownames(cor_numeric)[index[,1]], # get the row name 
                 variable_2 = colnames(cor_numeric)[index[,2]]) # get the column name

```
```{r}
# columns to remove from the correlation output
remove_cols <- c("open_acc","recoveries")
```


```{r}
##### Factor Screening #####
# get summary of the data
df_factor <- select_if(df, is.factor)

```


```{r}
# columns to remove from factor screening
remove_cols <- append(remove_cols, c("hardship_loan_status","hardship_flag",
                                     "pymnt_plan","issue_d","hardship_type",
                                     "hardship_reason", "debt_settlement_flag",
                                     "payment_plan_start_date",
                                     "settlement_date","debt_settlement_flag_date",
                                     "hardship_start_date","hardship_end_date",
                                     "next_pymnt_d","last_pymnt_d","last_credit_pull_d"))

# also remove dates as this should be time independent
remove_cols <- append(remove_cols, c("id","url","desc","disbursement_method",
                                     "sec_app_earliest_cr_line"))
```



```{r}
##### Final df to be used for further feature selection #####
df <- df[, -which(names(df) %in% remove_cols)]
summary(df)
```


In order to train the Boruta locally - we need to subset the data. 
```{r}
# subset the data
set.seed(7)
df <- df[1:50000,]

# make train and test split
sample <- caTools::sample.split(df$loan_status, SplitRatio = .70)
train <- subset(df, sample==TRUE)
test <- subset(df, sample==FALSE)
```

In the next step we run the Boruta algo.
```{r}
boruta_output <- Boruta(loan_status~., data = train, doTrace=2)
saveRDS(boruta_output, "2022_repo/outputs/boruta_output.rds")
```


Get and plot results 
```{r}
# Get significant variables including tentatives
boruta_signif <- getSelectedAttributes(boruta_output, withTentative = TRUE)
print(boruta_signif)

# Do a tentative rough fix
roughFixMod <- TentativeRoughFix(boruta_output)
boruta_signif <- getSelectedAttributes(roughFixMod)
print(boruta_signif)

# filter for the boruta variables
train <- train[, which(names(train) %in% append(boruta_signif,c("loan_status")))]
test <- test[,which(names(test) %in% append(boruta_signif, c("loan_status")))]

df_post_boruta <- rbind(test, train)
saveRDS(df_post_boruta, "df_post_boruta.rds")
```


```{r}
# Variable Importance Scores
imps <- attStats(roughFixMod)
imps2 = imps[imps$decision != 'Rejected', c('meanImp', 'decision')]
head(imps2[order(-imps2$meanImp), ])  # descending sort

# Plot variable importance
plot(boruta_output, cex.axis=.7, las=2, xlab="", main="Variable Importance") 

```


```{r}
# Train an rpart model and compute variable importance.
rPartMod <- train(loan_status ~ ., data=train, method="rpart")
rpartImp <- varImp(rPartMod)
print(rpartImp)
plot(rrfImp, top = 20, main='Variable Importance')

rrfMod <- train(loan_status ~ ., data=train, method="RRF")

rrfImp <- varImp(rrfMod, scale=F)

plot(rrfImp, top = 20, main='Variable Importance')
```
