==> Credit risk modelling use case/notebooks/04_data_exploration.Rmd <==
---
title: "Data exploration"
author: "41084.1_XAI_for_Finance"
output: html_document
---


This notebook contains the project's outputs with respect to WP4 and 5, specifically looking at data summary and visualization. The script provides various visualization techniques that represent meaningful summaries and features' interaction that risk experts would look at. 


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

libraries = c("readr", "corrplot", "RColorBrewer", "tidyverse", "kernlab", "e1071", "MLmetrics", "stargazer", "dplyr", "purrr", "xtable", "base", "ggplot2", "DescTools","stylo", "igraph", "xgboost", "DALEX", "DALEXtra", "stats", "ranger", "ALEPlot", "party", "lime", "MASS", "iml", "ROCR", "rpart", "e1071", "caret","MLmetrics", "Matrix", "Boruta", "readxl", "Metrics", "fastshap", "fscaret", "kknn", "knitr", "reshape", "shapper", "randomForest", "Hmisc", "networkD3", "emstreeR", "pdp", "pROC", "randomForestExplainer", "logistf")

lapply(libraries, function(x) if (!(x %in% installed.packages())) {
  install.packages(x)
})

lapply(libraries, library, quietly = TRUE, character.only = TRUE)

set.seed(7)
```


We are analyzing the sampled data after feature selection. For more details, check 2022_repo/01,02 and 03 notebooks. 

```{r}
df_post_boruta <- readRDS("~/GitHub/IWA_xai/2022_repo/Loan Performance Use Case/data/df_post_boruta.rds")
```


```{r}
df <- df_post_boruta
summary(df)
```

Grade and loan status. 
```{r}
ggplot(df, 
       aes(x = grade, 
           fill = loan_status)) + 
  geom_bar(position = "stack",
           color = "white") + 
  scale_fill_brewer(palette = "Pastel1")
```
```{r}
ggplot(df, 
       aes(x = grade, 
           fill = loan_status)) + 
  geom_bar(position = "fill") +
  scale_fill_brewer(palette = "Pastel1")
  
```

```{r, fig.width = 25, fig.height=25}
df %>%
  keep(is.numeric) %>% 
  gather() %>% 
  ggplot(aes(value)) +
    facet_wrap(~ key, scales = "free") +
    geom_histogram()
```



```{r, fig.width = 25, fig.height=25}
df %>%
  keep(is.factor) %>% 
  gather() %>% 
  ggplot(aes(value)) +
    facet_wrap(~ key, scales = "free") +
    geom_bar()

```


```{r}
numeric <- df %>%
  keep(is.numeric)
numeric$loan_status <- df$loan_status
for (i in numeric) {
  print(ggplot(numeric, aes(x = i, 
           y = loan_status,
           fill = loan_status)) +
  geom_density_ridges(alpha = 0.8) + 
  theme_ridges() + 
  labs(y = "Loan Status"))}
```


```{r}
factor <- df %>%
  keep(is.factor) 
for (i in factor) {
  print(ggplot(factor, aes(x = i,
           fill = loan_status)) +
  geom_bar(position = "stack",
           color = "white") +
    theme(axis.text.x = element_text(angle = 90)))}

```

```{r}
df <- dplyr::select(df, -3)
```

c(sum(numeric[i])), sd(numeric[i])
```{r}
numeric <- df %>%
  keep(is.numeric)

```


```{r}
colnames(df)
```


The scoring model we develop should mimic the Lending Club original scoring process. As such, all variables included in the data that indicate the status of the loan or borrowers' accounts after the loan has been approved, should be excluded from the analysis. Moreover, information on the grade and the interest rate assigned should also be excluded. 

```{r}
colnames(df)
post_approval = c("int_rate", "grade", "sub_grade", "out_prncp", "total_pymnt", "total_rec_int", "total_rec_late_fee", "collection_recovery_fee", "last_pymnt_amnt", "last_fico_range_high", "last_fico_range_low", "hardship_status", "settlement_status")
```


The final data set used for the analysis contains 59 variables across 50K loans. 
```{r}
df <- df[ , !(names(df) %in% post_approval)]
saveRDS(df, "df_post_data_explor.rds")
```
