==> Credit risk modelling use case/notebooks/05_data_summary_and_visualization.Rmd <==
---
title: "Data summary and visualization"
author: "41084.1_XAI_for_Finance"
output: html_document
---


This notebook contains the project's outputs with respect to WP4 and 5, specifically looking at data summary and visualization. The script provides various visualization techniques that represent meaningful summaries and features' interaction that risk experts would look at. 


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

libraries = c("readr", "corrplot", "RColorBrewer", "tidyverse", "kernlab", "e1071", "MLmetrics", "stargazer", "dplyr", "purrr", "xtable", "base", "ggplot2", "DescTools","stylo", "igraph", "xgboost", "DALEX", "DALEXtra", "stats", "ranger", "ALEPlot", "party", "lime", "MASS", "iml", "ROCR", "rpart", "e1071", "caret","MLmetrics", "Matrix", "Boruta", "Metrics", "fastshap", "fscaret", "kknn", "knitr", "reshape", "ggridges", "shapper", "scales", "randomForest", "readxl", "pastecs", "corrr", "PerformanceAnalytics", "Hmisc", "networkD3", "emstreeR", "pdp", "pROC", "ggcorrplot", "corrplot", "randomForestExplainer", "logistf")

lapply(libraries, function(x) if (!(x %in% installed.packages())) {
  install.packages(x)
})

lapply(libraries, library, quietly = TRUE, character.only = TRUE)

set.seed(7)
```

The app should contain a data summary and visualization tab where users can select a variable and obtain variety of useful information. Specifically, the user should obtain:
- a description of the variable (i.e. link abbriviation with full description of each variable)
- summary statistics of the variable (mean, median, etc.)
- visualization: if numeric:
  * histogram 
  * histogram (or density plot) per loan_status 
  * mean value per loan_status
  * mean value per year per loan_status
  * correlation with other numerical data
  * dependence between the var and the target (linear)


Final data used for the app: .... For more details, check 2022_repo/01,02,03 and 04 notebooks. 

```{r}
setwd("~/GitHub/IWA_xai")
df_post_data_explor <- readRDS("~/GitHub/IWA_xai/2022_repo/Loan Performance Use Case/data/df_post_data_explor.rds")
df <- df_post_data_explor
```

Import dictionary for variables included 
```{r}
 dictionary <- read_excel("~/GitHub/IWA_xai/2022_repo/Loan Performance Use Case/data/LCDataDictionary.xlsx")
```



## Visualization of variables: analysis and visualization

Indicate the variable to be summarized 
```{r}
var = "loan_amnt" #change the var name
```

Basic summary information
```{r}
dictionary$Description[dictionary$LoanStatNew == var]
stat.desc(df[var])
describe(df[var])
```


Distribution (histogram, box plot, cummulative plot)
```{r}
Desc(df[var], plotit = TRUE)
```

Histogram per loan status
```{r}
ggplot(df, 
       aes(x = df[,var])) +
  geom_histogram(fill = "darkseagreen1",
                 color = "grey") +
  facet_wrap(~loan_status, ncol = 1) 
```

Boxplot per loan status
```{r}
ggplot(df, aes(x = loan_status, 
                     y = df[,var],
       fill = loan_status)) +
  geom_boxplot(notch = TRUE,
               alpha = .5)

```

Boxplots per loan status (seperate graphs)
```{r}
ggplot(df, aes(x = loan_status, 
                     y = df[,var],
                     fill = loan_status)) +
  geom_boxplot() +
  facet_wrap(~loan_status, scale = "free") 
```


Boxplot + jitter

```{r}
ggplot(df, 
       aes(x = factor(loan_status,
                      labels = c("Paid",
                                 "Defaulted")), 
           y = df[,var], 
           color = loan_status)) +
  geom_boxplot(size=1,
               outlier.shape = 1,
               outlier.color = "black",
               outlier.size  = 3) +
  geom_jitter(alpha = 0.5, 
              width=.2) +
  theme_minimal() +
  theme(legend.position = "none") 
```


Density plots per loan status
```{r}
ggplot(df, 
       aes(x = df[,var], 
           y = loan_status,
           fill = loan_status)) +
  geom_density_ridges(alpha = 0.7) + 
  theme_ridges() +
  theme(legend.position = "none")
```

Mean plots per loan status 
```{r}
plotdata <- df %>%
  group_by(loan_status) %>%
  summarise(n = n(),
            mean = mean(df[,var]),
            sd = sd(df[,var]),
            se = sd/sqrt(n))


# Now let's plot these summaries 
ggplot(plotdata, 
       aes(x = loan_status, 
           y = mean, 
           group = 1)) +
  geom_point(size = 3) +
  geom_line() +
  geom_errorbar(aes(ymin = mean - se, 
                    ymax = mean + se), 
                width = .1)

```


```{r}
correlations <- df %>% 
  keep(is.numeric) %>%
  cor()
```


```{r, fig.width = 15, fig.height=15}
corrplot(correlations, method = "color")
```


```{r, fig.width = 25, fig.height=25}
# Check for significance of the correlations
numeric <- df %>% keep(is.numeric)
p_value_mat <- cor_pmat(numeric)
ggcorrplot(correlations, type = "lower", p.mat = p_value_mat) # all significant
```


Check preliminary dependence between var and loan_status 
```{r}
fit = glm(loan_status ~ loan_amnt, data = df, family = "binomial")
summary(fit)
```

 

# Qualitative features for var

```{r}
var = "term"
```


Bar plot per loan status 
```{r}
ggplot(df, 
       aes(x = df[,var], 
           fill = loan_status)) + 
  geom_bar(position = "dodge") + 
  labs (title = "Grouped barplot per loan status", x = colnames(df[var]))
```

```{r}
plotdata <- df %>%
  group_by(term, loan_status)  %>%
  dplyr::summarize(n = n()) %>% 
  mutate(pct = n/sum(n),
         lbl = scales::percent(pct))
```

Better representation 
```{r}
ggplot(plotdata,                                                         
       aes(x = term,
           y = pct,                                                     
           fill = loan_status)) + 
  geom_bar(stat = "identity",                                       
           position = "fill") +
  geom_text(aes(label = lbl),                                                                                       
            size = 3, 
            position = position_stack(vjust = 0.5)) +
  scale_fill_brewer(palette = "Set2") +
  theme_minimal()

```

Final stage: check the association between the qualitative feature and the target using Chi-Square test. The null hypothesis of the Chi-Square test is that there is no relationship whereas the alternative hypothesis assumes that there is an association between the two variables. 

```{r}
chisq.test(df[,var], df$loan_status)
```
